--- app/src/main/java/woowacourse/shopping/constant/Format.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.constant
-
-object Format {
-    fun formatPrice(price: Int): String = String.format("%,d원", price)
-}
