--- app/src/main/java/woowacourse/shopping/constant/ShoppingColor.kt
@@ -1,9 +1,0 @@
-package woowacourse.shopping.constant
-
-object ShoppingColor {
-    const val APP_BAR_COLOR = 0xFF555555
-    const val CART_ADD_BUTTON_COLOR = 0xFF14C3B1
-    const val PRODUCT_DETAIL_BACKGROUND_COLOR = 0xFFE0E0E0
-    const val CART_PAGE_BUTTON_ACTIVE_COLOR = 0xFF04C09E
-    const val CART_PAGE_BUTTON_INACTIVE_COLOR = 0xFFAAAAAA
-}
