--- app/src/main/java/woowacourse/shopping/domain/cart/CartItem.kt
@@ -1,11 +1,0 @@
-package woowacourse.shopping.domain.cart
-
-import woowacourse.shopping.domain.product.Product
-import java.util.UUID
-
-data class CartItem(
-    val id: String = UUID.randomUUID().toString(),
-    val product: Product,
-) {
-    fun isSameCartItem(targetCartItem: CartItem): Boolean = id == targetCartItem.id
-}
