+++ app/src/main/java/woowacourse/shopping/domain/cart/repository/CartRepository.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.domain.cart.repository
+
+import woowacourse.shopping.domain.cart.model.Cart
+import woowacourse.shopping.domain.cart.model.CartItem
+
+interface CartRepository {
+    suspend fun getCart(): Cart
+
+    suspend fun addCartItem(cartItem: CartItem)
+
+    suspend fun removeCartItem(cartItem: CartItem)
+}
