--- app/src/main/java/woowacourse/shopping/domain/product/ImageUrl.kt
@@ -1,10 +1,0 @@
-package woowacourse.shopping.domain.product
-
-@JvmInline
-value class ImageUrl(
-    val value: String,
-) {
-    init {
-        require(value.isNotBlank()) { "이미지 링크가 비어있습니다." }
-    }
-}
