--- app/src/main/java/woowacourse/shopping/domain/product/Price.kt
@@ -1,10 +1,0 @@
-package woowacourse.shopping.domain.product
-
-@JvmInline
-value class Price(
-    val value: Int,
-) {
-    init {
-        require(value >= 0) { "가격이 음수입니다." }
-    }
-}
