--- app/src/main/java/woowacourse/shopping/domain/product/Product.kt
@@ -1,10 +1,0 @@
-package woowacourse.shopping.domain.product
-
-import java.util.UUID
-
-data class Product(
-    val id: String = UUID.randomUUID().toString(),
-    val imageUrl: ImageUrl,
-    val name: ProductName,
-    val price: Price,
-)
