+++ app/src/main/java/woowacourse/shopping/domain/product/model/ImageUrl.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.domain.product.model
+
+@JvmInline
+value class ImageUrl(
+    val value: String,
+) {
+    init {
+        require(value.isNotBlank()) { "이미지 링크가 비어있습니다." }
+    }
+}
