+++ app/src/main/java/woowacourse/shopping/domain/product/model/Price.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.domain.product.model
+
+@JvmInline
+value class Price(
+    val value: Int,
+) {
+    init {
+        require(value >= 0) { "가격이 음수입니다." }
+    }
+}
