+++ app/src/main/java/woowacourse/shopping/domain/product/model/Product.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.domain.product.model
+
+import java.util.UUID
+
+data class Product(
+    val id: String = UUID.randomUUID().toString(),
+    val imageUrl: ImageUrl,
+    val name: ProductName,
+    val price: Price,
+)
