+++ app/src/main/java/woowacourse/shopping/domain/product/model/ProductName.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.domain.product.model
+
+@JvmInline
+value class ProductName(
+    val value: String,
+) {
+    init {
+        require(value.isNotBlank()) { "상품 이름이 비어있습니다." }
+    }
+}
