+++ app/src/main/java/woowacourse/shopping/features/cart/CartActivity.kt
@@ -1,0 +1,30 @@
+package woowacourse.shopping.features.cart
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.ui.Modifier
+
+class   CartActivity : ComponentActivity() {
+    private val viewModel: CartViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            Scaffold(
+                modifier = Modifier.fillMaxSize(),
+            ) { innerPadding ->
+                CartScreen(
+                    modifier = Modifier.padding(innerPadding),
+                    viewModel = viewModel
+                )
+            }
+        }
+    }
+}
