+++ app/src/main/java/woowacourse/shopping/features/constant/Format.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.features.constant
+
+object Format {
+    fun formatPrice(price: Int): String = String.format("%,d원", price)
+}
