+++ app/src/main/java/woowacourse/shopping/features/productList/ProductListActivity.kt
@@ -1,0 +1,43 @@
+package woowacourse.shopping.features.productList
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.ui.Modifier
+import woowacourse.shopping.features.cart.CartActivity
+import woowacourse.shopping.features.productDetail.ProductDetailActivity
+
+class ProductListActivity : ComponentActivity() {
+
+    private val viewModel: ProductListViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+                ProductListScreen(
+                    modifier = Modifier.padding(innerPadding),
+                    viewModel = viewModel,
+                    onCartClick = {
+                        val cartIntent = Intent(this, CartActivity::class.java)
+                        startActivity(cartIntent)
+                    },
+                    onProductClick = { product ->
+                        val detailIntent =
+                            Intent(this, ProductDetailActivity::class.java).apply {
+                                putExtra("PRODUCT_ID", product.id)
+                            }
+                        startActivity(detailIntent)
+                    },
+                )
+            }
+        }
+    }
+}
