--- app/src/main/java/woowacourse/shopping/repository/cart/CartRepository.kt
@@ -1,13 +1,0 @@
-package woowacourse.shopping.repository.cart
-
-import woowacourse.shopping.domain.cart.Cart
-import woowacourse.shopping.domain.cart.CartItem
-import woowacourse.shopping.domain.cart.CartItems
-
-interface CartRepository {
-    suspend fun getCart(): Cart
-
-    suspend fun addCartItem(cartItem: CartItem)
-
-    suspend fun removeCartItem(cartItem: CartItem)
-}
