--- app/src/main/java/woowacourse/shopping/repository/cart/CartRepositoryMockImpl.kt
@@ -1,19 +1,0 @@
-package woowacourse.shopping.repository.cart
-
-import woowacourse.shopping.domain.cart.Cart
-import woowacourse.shopping.domain.cart.CartItem
-import woowacourse.shopping.domain.cart.CartItems
-
-object CartRepositoryMockImpl : CartRepository {
-    private var cart: Cart = Cart(cartItems = CartItems(mutableListOf()))
-
-    override suspend fun getCart(): Cart = cart
-
-    override suspend fun addCartItem(cartItem: CartItem) {
-        cart = cart.addCartItem(cartItem)
-    }
-
-    override suspend fun removeCartItem(cartItem: CartItem) {
-        cart = cart.removeCartItem(cartItem)
-    }
-}
