+++ app/src/test/java/woowacourse/shopping/product/PriceTest.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.product
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import woowacourse.shopping.domain.product.Price
+import woowacourse.shopping.domain.product.model.Price
 
 class PriceTest {
     @Test
