+++ app/src/test/java/woowacourse/shopping/product/ProductNameTest.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.product
 
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
-import woowacourse.shopping.domain.product.ProductName
+import woowacourse.shopping.domain.product.model.ProductName
 
 class ProductNameTest {
     @Test
